revisions

+++ routes/web.php
@@ -4,9 +4,13 @@ use Illuminate\Support\Facades\Auth;
 
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\AuthController;
+use App\Http\Middleware\AdminMiddleware;
+use Illuminate\Support\Facades\Response;
 use App\Http\Controllers\EventsController;
 use App\Http\Controllers\SpeakersController;
 use App\Http\Controllers\DashboardController;
+use App\Http\Middleware\ParticipantMiddleware;
+use App\Http\Controllers\CooperativeController;
 use App\Http\Controllers\ParticipantController;
 use SebastianBergmann\CodeCoverage\Report\Html\Dashboard;
 
@@ -29,6 +33,10 @@ Route::post('/register', [AuthController::class, 'register']);
 Route::post('/logout', function () {Auth::logout();return redirect('/login');})->name('logout');
 
 //store
+Route::middleware([AdminMiddleware::class])->group(function () {
+
+Route::post('/cooperatives/{coop_id}/notify', [DashboardController::class, 'sendNotification'])->name('cooperatives.notify');
+
 Route::get('/Admin/Dashboard', [DashboardController::class, 'admin'])->name('adminDashboard');
 Route::get('/Admin/Register/Cooperatives', [DashboardController::class, 'register'])->name('adminregister');
 Route::post('/store-cooperative', [DashboardController::class, 'storeCooperative'])->name('admin.storeCooperative');
@@ -39,57 +47,35 @@ Route::get('/Admin/Cooperatives', [DashboardController::class, 'view'])->name('a
 Route::delete('/admin/cooperatives/{coop_id}', [DashboardController::class, 'destroy'])->name('cooperatives.destroy');
 
 //edit
-Route::get('/admin/cooperatives/{coop_id}/edit', [DashboardController::class, 'edit'])->name('cooperatives.edit');
-Route::put('/admin/cooperatives/{coop_id}', [DashboardController::class, 'update'])->name('cooperatives.update');
-Route::get('/cooperatives/{id}', [DashboardController::class, 'show'])->name('cooperatives.show');
+Route::get('/Admin/Cooperatives/Edit/{coop_id}', [DashboardController::class, 'edit'])->name('cooperatives.edit');
+Route::put('/Admin/Cooperatives/{coop_id}', [DashboardController::class, 'update'])->name('cooperatives.update');
+Route::get('/Admin/Cooperatives/View/{id}', [DashboardController::class, 'show'])->name('cooperatives.show');
 
 // store user
-Route::get('/user/register', [DashboardController::class, 'registerform'])->name('registerform');
+Route::get('/Admin/Users', [AuthController::class, 'index'])->name('users.index');
+Route::get('/Admin/User/Register', [DashboardController::class, 'registerform'])->name('registerform');
 Route::post('/user/register', [DashboardController::class, 'userregister'])->name('userregister');
-
-Route::get('/users', [AuthController::class, 'index'])->name('users.index');
 Route::delete('/users/{user_id}', [AuthController::class, 'destroy'])->name('users.destroy');
 
 // edit user
-Route::get('/user/{user_id}/edit', [AuthController::class, 'edit'])->name('user.edit');
-Route::put('/user/{user_id}', [AuthController::class, 'update'])->name('user.update');
+Route::get('/Admin/User/{user_id}/Edit', [AuthController::class, 'edit'])->name('user.edit');
+Route::put('/Admin/User/{user_id}', [AuthController::class, 'update'])->name('user.update');
 
 //participant admin crude
 
 Route::post('/participants/{participant}/approve', [ParticipantController::class, 'approve'])->name('participants.approve');
 
-Route::get('/participants', [ParticipantController::class, 'index'])->name('participants.index');
+Route::get('/Admin/Participants', [ParticipantController::class, 'index'])->name('participants.index');
 
-Route::get('/add/participant', [ParticipantController::class, 'participantadd'])->name('participantadd');
-Route::post('/add/participant', [ParticipantController::class, 'store'])->name('participant.add');
-Route::get('/participants/{participant_id}', [ParticipantController::class, 'show'])->name('participants.show');
+Route::get('/Admin/Add/Participant', [ParticipantController::class, 'participantadd'])->name('participantadd');
+Route::post('/Admin/Add/Participant', [ParticipantController::class, 'store'])->name('participant.add');
+Route::get('/Admin/Participant/{participant_id}', [ParticipantController::class, 'show'])->name('participants.show');
+Route::get('/Admin/Document/View/{participant_id?}', [ParticipantController::class, 'viewadminDocuments'])->name('admin.documents.view');
+Route::get('/Admin/Participants/{participant_id}/edit', [ParticipantController::class, 'edit'])->name('participants.edit');
+Route::put('/Participants/{participant_id}', [ParticipantController::class, 'update'])->name('participants.update');
+Route::delete('Admin/participants/{participant_id}', [ParticipantController::class, 'destroy'])->name('participants.destroy');
 
-Route::get('/participants/{participant_id}/edit', [ParticipantController::class, 'edit'])->name('participants.edit');
-Route::put('/participants/{participant_id}', [ParticipantController::class, 'update'])->name('participants.update');
-Route::delete('/participants/{participant_id}', [ParticipantController::class, 'destroy'])->name('participants.destroy');
-
-// participant dashboard
-
-Route::get('/participant/dashboard', [DashboardController::class, 'participant'])->name('participantDashboard');
-Route::get('/cooperativeprofile/{participant_id}/{coop_id}', [DashboardController::class, 'cooperativeprofile'])->name('cooperativeprofile');
-
-Route::get('/cooperativeprofile/{participant_id}/{coop_id}/edit', [DashboardController::class, 'editCooperativeProfile'])->name('cooperativeprofile.edit');
-Route::put('/cooperativeprofile/{participant_id}/{coop_id}', [DashboardController::class, 'updateCooperativeProfile'])->name('cooperativeprofile.update');
-
-
-Route::get('/participant/register', [DashboardController::class, 'participantregister'])->name('participant.register');
-Route::post('/participant/register', [DashboardController::class, 'store'])->name('participant.store');
-
-Route::get('/participant/upload/documents', [ParticipantController::class, 'documents'])->name('documents');
-Route::post('/participant/upload/documents', [ParticipantController::class, 'storeDocuments'])->name('documents.store');
-// Route::post('/participant/documents/upload', [ParticipantController::class, 'uploadDocuments'])->name('documents.upload');
-
-Route::get('/participant/documents', [ParticipantController::class, 'viewDocuments'])->name('documents.view');
-
-
-Route::get('/admin/documents/view/{participant_id?}', [ParticipantController::class, 'viewadminDocuments'])->name('admin.documents.view');
-
-Route::get('/admin/speakers', [SpeakersController::class, 'index'])->name('speakers.index');
+Route::get('/Admin/Speakers', [SpeakersController::class, 'index'])->name('speakers.index');
 
 Route::post('/speakers', [SpeakersController::class, 'store'])->name('speakers.store');
 Route::put('/speakers/{speaker_id}', [SpeakersController::class, 'update'])->name('speakers.update');
@@ -104,8 +90,79 @@ Route::put('/events/{event_id}', [EventsController::class, 'update'])->name('eve
 Route::get('/events/{event}/edit', [EventsController::class, 'edit'])->name('events.edit');
 Route::delete('/events/{event}', [EventsController::class, 'destroy'])->name('events.destroy');
 
+// profile admin
+Route::get('/admin/profile/edit', [AuthController::class, 'editProfile'])->name('profile.edit');
+Route::put('/adminp/rofile/edit', [AuthController::class, 'updateProfile'])->name('profile.update');
 
-Route::get('/participant/speakers', [ParticipantController::class, 'speakerlist'])->name('speakerlist');
+Route::get('participantQR/{id}', [DashboardController::class, 'showQR'])->name('participant.showQR');
 
-Route::get('/Schedules', [EventsController::class, 'schedule'])->name('schedule');
+Route::get('/download-qr2/{participant_id}', function ($participant_id) {
+    $qrCodeUrl = route('adminDashboard', ['participant_id' => $participant_id]);
+    $qrCodeImage = file_get_contents("https://api.qrserver.com/v1/create-qr-code/?data=" . urlencode($qrCodeUrl) . "&size=200x200");
+    return Response::make($qrCodeImage, 200, [
+        'Content-Type' => 'image/png',
+        'Content-Disposition' => 'attachment; filename="participant_' . $participant_id . '_qr_code.png"',
+    ]);
+})->name('download.qr2');
+
+});
+
+// participant user
+Route::middleware([ParticipantMiddleware::class])->group(function () {
+
+Route::get('/participant/dashboard', [DashboardController::class, 'participant'])->name('participantDashboard');
+Route::get('/cooperativeprofile/{participant_id}/{coop_id}', [DashboardController::class, 'cooperativeprofile'])->name('cooperativeprofile');
+
+Route::get('/cooperativeprofile/{participant_id}/{coop_id}/edit', [DashboardController::class, 'editCooperativeProfile'])->name('cooperativeprofile.edit');
+Route::put('/cooperativeprofile/{participant_id}/{coop_id}', [DashboardController::class, 'updateCooperativeProfile'])->name('cooperativeprofile.update');
+
+
+Route::get('/participant/register', [DashboardController::class, 'participantregister'])->name('participant.register');
+Route::post('/participant/register', [DashboardController::class, 'store'])->name('participant.store');
+
+Route::get('/participant/upload/documents', [ParticipantController::class, 'documents'])->name('documents');
+Route::post('/participant/upload/documents', [ParticipantController::class, 'storeDocuments'])->name('documents.store');
+
+// Route::get('/participant/documents', [ParticipantController::class, 'viewDocuments'])->name('documents.view');
+
+Route::get('/Participant/speakers', [ParticipantController::class, 'speakerlist'])->name('speakerlist');
+
+Route::get('/Participant/Event/Schedules', [EventsController::class, 'schedule'])->name('schedule');
+
+// profile participant
+Route::get('participant/profile/edit', [ParticipantController::class, 'editProfile'])->name('participant.profile.edit');
+Route::put('participant/profile/edit', [ParticipantController::class, 'updateProfile'])->name('participant.profile.update');
+
+Route::get('/download-qr/{participant_id}', function ($participant_id) {
+    $qrCodeUrl = route('adminDashboard', ['participant_id' => $participant_id]);
+    $qrCodeImage = file_get_contents("https://api.qrserver.com/v1/create-qr-code/?data=" . urlencode($qrCodeUrl) . "&size=200x200");
+    return Response::make($qrCodeImage, 200, [
+        'Content-Type' => 'image/png',
+        'Content-Disposition' => 'attachment; filename="participant_' . $participant_id . '_qr_code.png"',
+    ]);
+})->name('download.qr');
+
+// participant crud coop
+
+Route::get('/Cooperative/Participants', [CooperativeController::class, 'index'])->name('coop.index');
+
+Route::get('/Cooperative/Add/Participant', [CooperativeController::class, 'coopparticipantadd'])->name('coopparticipantadd');
+Route::post('/Cooperative/Add/Participant', [CooperativeController::class, 'store'])->name('coopparticipant.add');
+Route::get('/Cooperative/Participant/{participant_id}', [CooperativeController::class, 'show'])->name('coop.participants.show');
+
+// Route::get('/Admin/Document/View/{participant_id?}', [ParticipantController::class, 'viewadminDocuments'])->name('admin.documents.view');
+
+Route::get('/Cooperatives/Participants/{participant_id}/edit', [CooperativeController::class, 'edit'])->name('coop.participants.edit');
+Route::put('Coop/Participants/{participant_id}', [CooperativeController::class, 'update'])->name('coop.participants.update');
+Route::delete('Cooperative/delete/participants/{participant_id}', [CooperativeController::class, 'destroy'])->name('coop.participants.destroy');
+
+Route::get('/Cooperative/upload/documents', [CooperativeController::class, 'documents'])->name('documents');
+Route::post('/Cooperative/upload/documents', [CooperativeController::class, 'storeDocuments'])->name('documents.store');
+
+
+Route::get('/Cooperative/documents', [CooperativeController::class, 'viewDocuments'])->name('documents.view');
+
+});
+
+
